requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
altgraph==0.17
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.4
emoji==1.2.0
env==0.1.0
environs==9.3.1
future==0.18.2
idna==2.10
marshmallow==3.10.0
multidict==5.1.0
pefile==2019.4.18
pyarmor==6.6.1
pycparser==2.20
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyTelegramBotAPI==3.7.6
python-dotenv==0.15.0
python-telegram-bot==13.2
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.25.1
six==1.15.0
soupsieve==2.2
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
yarl==1.6.3